+++ src/gromacs/analysisdata/modules/lifetime.cpp
@@ -133,7 +133,7 @@ void AnalysisDataLifetimeModule::dataStarted(AbstractAnalysisData* data)
 {
     impl_->currentLifetimes_.reserve(data->dataSetCount());
     impl_->lifetimeHistograms_.reserve(data->dataSetCount());
-    for (int i = 0; i < data->dataSetCount(); ++i)
+    for (size_t i = 0; i < data->dataSetCount(); ++i)
     {
         impl_->currentLifetimes_.emplace_back(data->columnCount(i), 0);
         impl_->lifetimeHistograms_.emplace_back();
@@ -235,7 +235,7 @@ void AnalysisDataLifetimeModule::dataFinished()
     for (histogram = impl_->lifetimeHistograms_.begin(); histogram != impl_->lifetimeHistograms_.end();
          ++histogram, ++column)
     {
-        int                                     row = 0;
+        size_t                                  row = 0;
         Impl::LifetimeHistogram::const_iterator i;
         for (i = histogram->begin(); i != histogram->end(); ++i, ++row)
         {
